agents/aws_agent/requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
mcp
